Updated change-host, now with incasso values
1. Updated change-host, now with incasso values
2. And placed db-stuff in its own file
